+++ static/form/style.css
@@ -1,4 +1,3 @@
-/* .scenario_area input[type="text"], */
 .scenario_area {
     width:100%;
     height:50vh;
@@ -9,7 +8,6 @@
     font-size: 16px;
     font-family: "Hiragino Maru Gothic ProN";
 }
-/* .scenario_area input[type="text"]:focus, */
 .scenario_area:focus {
     box-shadow: 0 0 7px #1abc9c;
     border: 1px solid #1abc9c;
@@ -26,7 +24,7 @@
   border-bottom-style: none;
 }
 
-.square_btn {
+.submitbtn {
     width: 33vw;
     height: 80px;
     display: inline-block;
@@ -41,19 +39,12 @@
     font-family: "Hiragino Maru Gothic ProN";
 }
 
-::-webkit-input-placeholder {  color: #000;}
-:-moz-placeholder { color: #000; }
-
-.square_btn:hover {
+.submitbtn:hover {
     background: #1abc9c;
     color: white;
 }
-/*
-.form_area{
-    margin: 10px 727px 10px 30px;
-} */
 
-.charanameInp{
+.name{
   width: 100%;
   vertical-align: top;
   border:1px #1abc9c solid;
@@ -61,7 +52,7 @@
   font-family: 'Hiragino Maru Gothic ProN';
 }
 
-.select{
+.role{
     width: 80%;
     height: 25px;
     -moz-appearance: none;
@@ -89,7 +80,7 @@
     vertical-align: top;
 }
 
-.btn {
+.appendBtn {
     width: 100%;
     height: 27px;
     border-radius: 3px;
@@ -102,11 +93,11 @@
     transition: .4s;
     vertical-align: top;
 }
-.btn:hover {
+.appendBtn:hover {
     background: #1abc9c;
     color: white;
 }
-
+/*
 .tab{
   overflow:hidden; margin: 0
 }
@@ -189,4 +180,4 @@
 @keyframes rotate {
     0%    { transform: rotate(0deg); }
     100%  { transform: rotate(360deg); }
-}
+} */
